+++ R/html_document.R
@@ -85,7 +85,8 @@
 #'  default definition or R Markdown. See the \code{\link{rmarkdown_format}} for
 #'  additional details.
 #'@param pandoc_args Additional command line options to pass to pandoc
-#'@param extra_dependencies,... Additional function arguments to pass to the
+#'@param extra_dependencies,allow_uptree_lib_dir,... Additional function
+#'  arguments to pass to the
 #'  base R Markdown HTML output formatter \code{\link{html_document_base}}
 #'@return R Markdown output format to pass to \code{\link{render}}
 #'
@@ -202,6 +203,39 @@
 #'  Due to the above restrictions, you might consider using the \code{includes}
 #'  parameter as an alternative to providing a fully custom template.
 #'
+#'@section Directory structure:
+#'
+#'  By default `html_document` and related HTML document types put dependency
+#'  files into the main output directory or a subdirectory specified by the
+#'  `lib_dir` parameter:
+#'  If `lib_dir` is not a direct descendant of the
+#'  main output directory, `render()` will throw and error with the
+#'  message "The path <lib_dir> does not appear to be a descendant of
+#'  <output_dir>".
+#'
+#'  Sometimes it is useful to have a directory tree where the different
+#'  HTML documents are in their own subdirectories and the
+#'  dependencies are in a common directory at the root of the site.
+#'
+#'  * main_dir/
+#'    * lib/
+#'      * dependencies go here
+#'    * index.Rmd
+#'    * node-01/
+#'      * index.Rmd
+#'    * node-02/
+#'      * index.Rmd
+#'
+#'  One way to achieve this is with the `render_site` command, but knitting
+#'  individual documents in subdirectories (such as with the RStudio "knit"
+#'  button) will result in errors.
+#'
+#'  It is possible to achieve this directory structure by setting the
+#'  `allow_uptree_lib_dir` parameter to `yes` or `true` in the
+#'  `output/html_document` section of the YAML header
+#'  and set `lib_dir` to a relative path, such as `../lib` or `../site_lib`
+#'  in `node-01/index.Rmd` and `node-02/index.Rmd` in the example above.
+#'
 #' @examples
 #' \dontrun{
 #' library(rmarkdown)
@@ -238,6 +272,7 @@ html_document <- function(toc = FALSE,
                           lib_dir = NULL,
                           md_extensions = NULL,
                           pandoc_args = NULL,
+                          allow_uptree_lib_dir = FALSE,
                           ...) {
 
   # build pandoc args
@@ -488,6 +523,7 @@ html_document <- function(toc = FALSE,
                                      pandoc_args = pandoc_args,
                                      extra_dependencies = extra_dependencies,
                                      css = css,
+                                     allow_uptree_lib_dir = allow_uptree_lib_dir,
                                      ...)
   )
 }