+++ R/output_format.R
@@ -56,6 +56,8 @@
 #'   should return a named list of files w/ \code{name} and \code{content} for
 #'   each file.
 #' @param base_format An optional format to extend.
+#' @param allow_uptree_lib_dir Allow `lib_dir` output parameter not to
+#'   be a descendent of the output directory.
 #' @return An R Markdown output format definition that can be passed to
 #'   \code{\link{render}}.
 #' @seealso \link{render}, \link{knitr_options}, \link{pandoc_options}
@@ -77,7 +79,8 @@ output_format <- function(knitr,
                           post_processor = NULL,
                           on_exit = NULL,
                           file_scope = NULL,
-                          base_format = NULL) {
+                          base_format = NULL,
+                          allow_uptree_lib_dir = FALSE) {
 
   format <- list(
     knitr = knitr,
@@ -91,7 +94,8 @@ output_format <- function(knitr,
     intermediates_generator = intermediates_generator,
     post_processor = post_processor,
     file_scope = file_scope,
-    on_exit = on_exit
+    on_exit = on_exit,
+    allow_uptree_lib_dir = allow_uptree_lib_dir
   )
 
   class(format) <- "rmarkdown_output_format"
@@ -166,6 +170,8 @@ merge_output_formats <- function(base,
     pandoc = merge_pandoc_options(base$pandoc, overlay$pandoc),
     keep_md =
       merge_scalar(base$keep_md, overlay$keep_md),
+    allow_uptree_lib_dir =
+      merge_scalar(base$allow_uptree_lib_dir, overlay$allow_uptree_lib_dir),
     clean_supporting =
       merge_scalar(base$clean_supporting, overlay$clean_supporting),
     df_print =